% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNR-package.R
\docType{package}
\name{BNR}
\alias{BNR}
\alias{BNR-package}
\title{BNR: Bivariate Normal Regression}
\description{
This package performs estimation and inference for the parameters of a
bivariate normal regression model. Estimation is performed using \code{\link{fit.bnr}}.
Inference on regression parameters for the target outcome is performed using
\code{\link{test.bnr}}.
}
\author{
Zachary R. McCaw
}
