% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fitting.R
\name{fit.bnr}
\alias{fit.bnr}
\title{Fit Bivariate Normal Regression Model.}
\usage{
fit.bnr(y1, y2, X, sig = 0.05)
}
\arguments{
\item{y1}{First outcome vector.}

\item{y2}{Second outcome vector.}

\item{X}{Target model matrix.}

\item{sig}{Significance level.}
}
\value{
An object of class 'bnr' with slots containing the estimated regression
 coefficients, the target-surrogate covariance matrix, the information matrices
 for regression parameters, and the residuals.
}
\description{
Estimation procedure for bivariate normal regression models.
}
