% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Testing.R
\name{test.bnr}
\alias{test.bnr}
\title{Test Bivariate Normal Regression Model.}
\usage{
test.bnr(y1, y2, X, L, test = "Wald", both = TRUE)
}
\arguments{
\item{y1}{First outcome vector.}

\item{y2}{Second outcome vector.}

\item{X}{Model matrix.}

\item{L}{Logical vector, with as many entires as columns in the target model
matrix, indicating which columns have coefficient zero under the null.}

\item{test}{Either Score or Wald. Only Wald is available with unilateral.}

\item{both}{If TRUE, performs the bilateral test.}
}
\value{
A numeric vector containing the test statistic, the degrees of
  freedom, and a p-value.
}
\description{
Inference procedure for bivariate normal regression models. Performs bilateral
tests that a subset of the regression parameters are zero for both outcomes,
and unilateral tests that a subset of the regression parameters are zero for
the first outcome only.
}
