% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wald_2.R
\name{Wald2}
\alias{Wald2}
\title{Bilateral Wald Test via Least Squares.}
\usage{
Wald2(y1, y2, X, L)
}
\arguments{
\item{y1}{First outcome vector.}

\item{y2}{Second outcome vector.}

\item{X}{Model matrix.}

\item{L}{Logical vector, with as many entires as columns in the design
matrix, indicating which columns have coefficient zero under the null.}
}
\value{
A numeric vector containing the score statistic, the degrees of
  freedom, and a p-value.
}
\description{
Performs a Wald test of the null hypothesis that a subset of the regression
parameters are zero for both outcomes.
}
